refactor(tangle-dapp): move polkadot api to shared ui lib
