+++ apps/tangle-dapp/context/PolkadotApiContext.tsx
@@ -5,6 +5,10 @@ import { ApiPromise } from '@polkadot/api/promise';
 import { ApiRx } from '@polkadot/api/rx';
 import type { Maybe } from '@webb-tools/dapp-types/utils/types';
 import useNetworkStore from '@webb-tools/tangle-shared-ui/context/useNetworkStore';
+import {
+  getApiPromise,
+  getApiRx,
+} from '@webb-tools/tangle-shared-ui/utils/polkadot/api';
 import noop from 'lodash/noop';
 import {
   createContext,
@@ -17,8 +21,6 @@ import {
 } from 'react';
 import useSWRImmutable from 'swr/immutable';
 import { Prettify } from 'viem/chains';
-
-import { getApiPromise, getApiRx } from '../utils/polkadot';
 
 export type PolkadotApiContextProps = Prettify<{
   apiPromise: ApiPromise;
