+++ apps/tangle-dapp/data/KeyStats/useIdealStakePercentage.ts
@@ -3,10 +3,10 @@
 import { BN_ZERO } from '@polkadot/util';
 import useNetworkStore from '@webb-tools/tangle-shared-ui/context/useNetworkStore';
 import ensureError from '@webb-tools/tangle-shared-ui/utils/ensureError';
+import { getApiPromise } from '@webb-tools/tangle-shared-ui/utils/polkadot/api';
 import { useEffect, useState } from 'react';
 
 import { calculateInflation } from '../../utils';
-import { getApiPromise } from '../../utils/polkadot';
 
 export default function useIdealStakedPercentage(
   defaultValue: { value1: number | null } = { value1: null },
