+++ apps/tangle-dapp/data/KeyStats/useInflationPercentage.ts
@@ -3,12 +3,15 @@
 import { BN_ZERO } from '@polkadot/util';
 import useNetworkStore from '@webb-tools/tangle-shared-ui/context/useNetworkStore';
 import ensureError from '@webb-tools/tangle-shared-ui/utils/ensureError';
+import {
+  getApiPromise,
+  getApiRx,
+} from '@webb-tools/tangle-shared-ui/utils/polkadot/api';
 import { useEffect, useState } from 'react';
 import { Subscription } from 'rxjs';
 
 import useFormatReturnType from '../../hooks/useFormatReturnType';
 import { calculateInflation } from '../../utils';
-import { getApiPromise, getApiRx } from '../../utils/polkadot';
 
 export default function useInflationPercentage(
   defaultValue: { value1: number | null; value2: number | null } = {
