+++ apps/tangle-dapp/data/NominatorStats/useTotalStakedAmountSubscription.ts
@@ -4,12 +4,12 @@ import { u128 } from '@polkadot/types';
 import { BN } from '@polkadot/util';
 import { WebbError, WebbErrorCodes } from '@webb-tools/dapp-types/WebbError';
 import useNetworkStore from '@webb-tools/tangle-shared-ui/context/useNetworkStore';
+import { getApiRx } from '@webb-tools/tangle-shared-ui/utils/polkadot/api';
 import { useEffect, useState } from 'react';
 import { type Subscription } from 'rxjs';
 
 import useFormatReturnType from '../../hooks/useFormatReturnType';
 import useSubstrateAddress from '../../hooks/useSubstrateAddress';
-import { getApiRx } from '../../utils/polkadot';
 
 export default function useTotalStakedAmountSubscription(
   defaultValue: { value1: BN | null } = { value1: null },
