+++ apps/tangle-dapp/data/liquidStaking/fetchHelpers.ts
@@ -2,6 +2,7 @@ import { Option } from '@polkadot/types';
 import { AccountId32, Perbill } from '@polkadot/types/interfaces';
 import { SpStakingPagedExposureMetadata } from '@polkadot/types/lookup';
 import { BN, BN_ZERO } from '@polkadot/util';
+import { getApiPromise } from '@webb-tools/tangle-shared-ui/utils/polkadot/api';
 
 import {
   PalletDappStakingV3ContractStakeAmount,
@@ -9,7 +10,6 @@ import {
 } from '../../types/astarDappStaking';
 import {
   extractDataFromIdentityInfo,
-  getApiPromise,
   IdentityDataType,
 } from '../../utils/polkadot';
