+++ apps/tangle-dapp/hooks/useApi.ts
@@ -3,7 +3,7 @@ import useNetworkStore from '@webb-tools/tangle-shared-ui/context/useNetworkStor
 import ensureError from '@webb-tools/tangle-shared-ui/utils/ensureError';
 import { useCallback, useEffect, useState } from 'react';
 
-import { getApiPromise } from '../utils/polkadot';
+import { getApiPromise } from '@webb-tools/tangle-shared-ui/utils/polkadot/api';
 import usePromise from './usePromise';
 
 export type ApiFetcher<T> = (api: ApiPromise) => Promise<T> | T;
