+++ apps/tangle-dapp/hooks/useNetworkSwitcher.ts
@@ -13,6 +13,7 @@ import useLocalStorage, {
   LocalStorageKey,
 } from '@webb-tools/tangle-shared-ui/hooks/useLocalStorage';
 import ensureError from '@webb-tools/tangle-shared-ui/utils/ensureError';
+import { getApiPromise } from '@webb-tools/tangle-shared-ui/utils/polkadot/api';
 import testRpcEndpointConnection from '@webb-tools/tangle-shared-ui/utils/testRpcEndpointConnection';
 import { calculateTypedChainId, ChainType } from '@webb-tools/utils';
 import { notificationApi } from '@webb-tools/webb-ui-components';
@@ -22,8 +23,6 @@ import {
 } from '@webb-tools/webb-ui-components/constants/networks';
 import { useCallback, useEffect, useState } from 'react';
 import { createPublicClient, fallback, http, webSocket } from 'viem';
-
-import { getApiPromise } from '../utils/polkadot';
 
 const useNetworkSwitcher = () => {
   const { switchChain, activeWallet } = useWebContext();
