+++ apps/tangle-dapp/hooks/useSubstrateTx.ts
@@ -4,6 +4,7 @@ import { ISubmittableResult } from '@polkadot/types/types';
 import { PromiseOrT } from '@webb-tools/abstract-api-provider';
 import useNetworkStore from '@webb-tools/tangle-shared-ui/context/useNetworkStore';
 import ensureError from '@webb-tools/tangle-shared-ui/utils/ensureError';
+import { getApiPromise } from '@webb-tools/tangle-shared-ui/utils/polkadot/api';
 import assert from 'assert';
 import { useCallback, useEffect, useState } from 'react';
 import { Hash } from 'viem';
@@ -11,7 +12,6 @@ import { Hash } from 'viem';
 import { TxName } from '../constants';
 import { GetSuccessMessageFunction } from '../types';
 import extractErrorFromTxStatus from '../utils/extractErrorFromStatus';
-import { getApiPromise } from '../utils/polkadot';
 import useActiveAccountAddress from './useActiveAccountAddress';
 import useAgnosticAccountInfo from './useAgnosticAccountInfo';
 import useIsMountedRef from './useIsMountedRef';
